Add -Wl,--soname=libns3.so to link flags to allow dlopen hacks
Mathieu's explanation:
"""
In python, I want to load the ns3 python bindings without having to
install libns3.so in /usr and without having to configure
LD_LIBRARY_PATH. So, I need to dlopen (RTLD_GLOBAL) libns3.so before
doing imp.find_module ('ns3') to ensure that python finds the libns3.so
library. To do the dlopen, I need to give a full path to libns3.so but,
if you do this without a soname, the dynamic loader does not know about
libns3.so: it knows only about /foo/bar/libns3.so.
"""